Fix bug where we were preventing paperclip from calling a method internally
We monkey-patch `Paperclip::Attachment#url` to prevent developers from using it in development,
since it won't work in production.  (`expiring_url` should be used instead.)

However, in development mode, the url method is called by default whenever the path option is
interpolated.  This commit circumvents that by avoind the need to interpolate the url in
order to interpolate the path.